More refactor and performance fixes
