+++ maps/infestation_islands/core.lua
@@ -1,8 +1,10 @@
 local Public = require 'maps.infestation_islands.table'
-Public.func = require 'maps.infestation_islands.func'
+Public.functions = require 'maps.infestation_islands.functions'
+Public.callbacks = require 'maps.infestation_islands.callbacks'
 Public.basic_markets = require 'maps.infestation_islands.basic_markets'
 Public.island_settings = require 'maps.infestation_islands.island_settings'
 Public.loot = require 'maps.infestation_islands.loot'
 Public.buried_biters = require 'maps.infestation_islands.buried_biters'
+Public.commands = require 'maps.infestation_islands.commands'
 
 return Public
