+++ maps/infestation_islands/island_settings.lua
@@ -1,877 +1,877 @@
 local map_gen_settings =
 {
-	autoplace_controls =
-	{
-		['coal'] = { frequency = 0, size = 0, richness = 0 },
-		['stone'] = { frequency = 0, size = 0, richness = 0 },
-		['copper-ore'] = { frequency = 0, size = 0, richness = 0 },
-		['iron-ore'] = { frequency = 0, size = 0, richness = 0 },
-		['uranium-ore'] = { frequency = 0, size = 0, richness = 0 },
-		['crude-oil'] = { frequency = 0, size = 0, richness = 0 },
-		['trees'] = { frequency = 50, size = 0.1, richness = 6 * 0.1 },
-		['enemy-base'] = { frequency = 'none', size = 'none', richness = 'none' },
-		nauvis_cliff =
-		{
-			frequency = 0,
-			richness = 0,
-			size = 0
-		},
-		rocks =
-		{
-			frequency = 1,
-			richness = 1,
-			size = 1
-		},
-		starting_area_moisture =
-		{
-			frequency = 1,
-			richness = 1,
-			size = 1
-		},
-		gleba_plants =
-		{
-			frequency = 1,
-			richness = 1,
-			size = 1
-		},
-		gleba_stone =
-		{
-			frequency = 1,
-			richness = 1,
-			size = 1
-		},
-		gleba_water =
-		{
-			frequency = 1,
-			richness = 1,
-			size = 1
-		}
-	},
-	autoplace_settings =
-	{
-		decorative =
-		{
-			settings =
-			{
-				["barnacles-decal"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["black-sceptre"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["blood-grape"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["blood-grape-vibrant"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				brambles =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["brown-cup"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["coral-land"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["coral-stunted"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["coral-stunted-grey"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["coral-water"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["cracked-mud-decal"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["cream-nerve-roots-veins-dense"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["cream-nerve-roots-veins-sparse"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["curly-roots-orange"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["dark-mud-decal"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["fuchsia-pita"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["green-bush-mini"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["green-carpet-grass"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["green-croton"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["green-cup"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["green-hairy-grass"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["green-lettuce-lichen-1x1"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["green-lettuce-lichen-3x3"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["green-lettuce-lichen-6x6"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["green-lettuce-lichen-water-1x1"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["green-lettuce-lichen-water-3x3"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["green-lettuce-lichen-water-6x6"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["green-pita"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["green-pita-mini"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["grey-cracked-mud-decal"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["honeycomb-fungus"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["honeycomb-fungus-1x1"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["honeycomb-fungus-decayed"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["knobbly-roots"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["knobbly-roots-orange"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["lichen-decal"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["light-mud-decal"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["matches-small"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				mycelium =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["pale-lettuce-lichen-1x1"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["pale-lettuce-lichen-3x3"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["pale-lettuce-lichen-6x6"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["pale-lettuce-lichen-cups-1x1"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["pale-lettuce-lichen-cups-3x3"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["pale-lettuce-lichen-cups-6x6"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["pale-lettuce-lichen-water-1x1"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["pale-lettuce-lichen-water-3x3"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["pale-lettuce-lichen-water-6x6"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["pink-lichen-decal"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["pink-phalanges"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["polycephalum-balloon"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["polycephalum-slime"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["purple-nerve-roots-veins-dense"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["purple-nerve-roots-veins-sparse"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["red-desert-bush"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["red-lichen-decal"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["red-nerve-roots-veins-dense"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["red-nerve-roots-veins-sparse"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["red-pita"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["shroom-decal"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["solo-barnacle"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["split-gill-1x1"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["split-gill-2x2"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["split-gill-dying-1x1"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["split-gill-dying-2x2"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["split-gill-red-1x1"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["split-gill-red-2x2"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				veins =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["veins-small"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["white-carpet-grass"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["white-desert-bush"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["wispy-lichen"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["yellow-coral"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["yellow-lettuce-lichen-1x1"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["yellow-lettuce-lichen-3x3"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["yellow-lettuce-lichen-6x6"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["yellow-lettuce-lichen-cups-1x1"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["yellow-lettuce-lichen-cups-3x3"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["yellow-lettuce-lichen-cups-6x6"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				}
-			},
-			treat_missing_as_default = true
-		},
-		entity =
-		{
-			settings =
-			{
-				["big-rock"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["big-sand-rock"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				coal =
-				{
-					frequency = 0,
-					richness = 0,
-					size = 0
-				},
-				["copper-ore"] =
-				{
-					frequency = 0,
-					richness = 0,
-					size = 0
-				},
-				["crude-oil"] =
-				{
-					frequency = 0,
-					richness = 0,
-					size = 0
-				},
-				fish =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["huge-rock"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["iron-ore"] =
-				{
-					frequency = 0,
-					richness = 0,
-					size = 0
-				},
-				stone =
-				{
-					frequency = 0,
-					richness = 0,
-					size = 0
-				},
-				["uranium-ore"] =
-				{
-					frequency = 0,
-					richness = 0,
-					size = 0
-				}
-			},
-			treat_missing_as_default = true
-		},
-		tile =
-		{
-			settings =
-			{
-				["gleba-deep-lake"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["highland-dark-rock"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["highland-dark-rock-2"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["highland-yellow-rock"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["lowland-brown-blubber"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["lowland-cream-cauliflower"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["lowland-cream-cauliflower-2"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["lowland-cream-red"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["lowland-dead-skin"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["lowland-dead-skin-2"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["lowland-olive-blubber"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["lowland-olive-blubber-2"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["lowland-olive-blubber-3"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["lowland-pale-green"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["lowland-red-infection"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["lowland-red-vein"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["lowland-red-vein-2"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["lowland-red-vein-3"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["lowland-red-vein-4"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["lowland-red-vein-dead"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["midland-cracked-lichen"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["midland-cracked-lichen-dark"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["midland-cracked-lichen-dull"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["midland-turquoise-bark"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["midland-turquoise-bark-2"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["midland-yellow-crust"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["midland-yellow-crust-2"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["midland-yellow-crust-3"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["midland-yellow-crust-4"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["natural-jellynut-soil"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["natural-yumako-soil"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["pit-rock"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["wetland-blue-slime"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["wetland-dead-skin"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["wetland-green-slime"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["wetland-jellynut"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["wetland-light-dead-skin"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["wetland-light-green-slime"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["wetland-pink-tentacle"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["wetland-red-tentacle"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				},
-				["wetland-yumako"] =
-				{
-					frequency = 1,
-					richness = 1,
-					size = 1
-				}
-			},
-			treat_missing_as_default = true
-		}
-	},
-	cliff_settings =
-	{
-		cliff_elevation_0 = 0,
-		cliff_elevation_interval = 0,
-		cliff_smoothing = 0,
-		control = "nauvis_cliff",
-		name = "cliff",
-		richness = 1
-	},
-	default_enable_all_autoplace_controls = false,
-	no_enemies_mode = false,
-	peaceful_mode = false,
-	property_expression_names = {},
-	starting_area = 10,
-	starting_points =
-	{
-		{
-			x = 0,
-			y = 0
-		}
-	},
+    autoplace_controls =
+    {
+        ['coal'] = { frequency = 0, size = 0, richness = 0 },
+        ['stone'] = { frequency = 0, size = 0, richness = 0 },
+        ['copper-ore'] = { frequency = 0, size = 0, richness = 0 },
+        ['iron-ore'] = { frequency = 0, size = 0, richness = 0 },
+        ['uranium-ore'] = { frequency = 0, size = 0, richness = 0 },
+        ['crude-oil'] = { frequency = 0, size = 0, richness = 0 },
+        ['trees'] = { frequency = 50, size = 0.1, richness = 6 * 0.1 },
+        ['enemy-base'] = { frequency = 'none', size = 'none', richness = 'none' },
+        nauvis_cliff =
+        {
+            frequency = 0,
+            richness = 0,
+            size = 0
+        },
+        rocks =
+        {
+            frequency = 1,
+            richness = 1,
+            size = 1
+        },
+        starting_area_moisture =
+        {
+            frequency = 1,
+            richness = 1,
+            size = 1
+        },
+        gleba_plants =
+        {
+            frequency = 1,
+            richness = 1,
+            size = 1
+        },
+        gleba_stone =
+        {
+            frequency = 1,
+            richness = 1,
+            size = 1
+        },
+        gleba_water =
+        {
+            frequency = 1,
+            richness = 1,
+            size = 1
+        }
+    },
+    autoplace_settings =
+    {
+        decorative =
+        {
+            settings =
+            {
+                ["barnacles-decal"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["black-sceptre"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["blood-grape"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["blood-grape-vibrant"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                brambles =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["brown-cup"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["coral-land"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["coral-stunted"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["coral-stunted-grey"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["coral-water"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["cracked-mud-decal"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["cream-nerve-roots-veins-dense"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["cream-nerve-roots-veins-sparse"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["curly-roots-orange"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["dark-mud-decal"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["fuchsia-pita"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["green-bush-mini"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["green-carpet-grass"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["green-croton"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["green-cup"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["green-hairy-grass"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["green-lettuce-lichen-1x1"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["green-lettuce-lichen-3x3"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["green-lettuce-lichen-6x6"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["green-lettuce-lichen-water-1x1"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["green-lettuce-lichen-water-3x3"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["green-lettuce-lichen-water-6x6"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["green-pita"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["green-pita-mini"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["grey-cracked-mud-decal"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["honeycomb-fungus"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["honeycomb-fungus-1x1"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["honeycomb-fungus-decayed"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["knobbly-roots"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["knobbly-roots-orange"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["lichen-decal"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["light-mud-decal"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["matches-small"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                mycelium =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["pale-lettuce-lichen-1x1"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["pale-lettuce-lichen-3x3"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["pale-lettuce-lichen-6x6"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["pale-lettuce-lichen-cups-1x1"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["pale-lettuce-lichen-cups-3x3"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["pale-lettuce-lichen-cups-6x6"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["pale-lettuce-lichen-water-1x1"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["pale-lettuce-lichen-water-3x3"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["pale-lettuce-lichen-water-6x6"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["pink-lichen-decal"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["pink-phalanges"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["polycephalum-balloon"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["polycephalum-slime"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["purple-nerve-roots-veins-dense"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["purple-nerve-roots-veins-sparse"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["red-desert-bush"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["red-lichen-decal"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["red-nerve-roots-veins-dense"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["red-nerve-roots-veins-sparse"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["red-pita"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["shroom-decal"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["solo-barnacle"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["split-gill-1x1"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["split-gill-2x2"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["split-gill-dying-1x1"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["split-gill-dying-2x2"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["split-gill-red-1x1"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["split-gill-red-2x2"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                veins =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["veins-small"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["white-carpet-grass"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["white-desert-bush"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["wispy-lichen"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["yellow-coral"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["yellow-lettuce-lichen-1x1"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["yellow-lettuce-lichen-3x3"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["yellow-lettuce-lichen-6x6"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["yellow-lettuce-lichen-cups-1x1"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["yellow-lettuce-lichen-cups-3x3"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["yellow-lettuce-lichen-cups-6x6"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                }
+            },
+            treat_missing_as_default = true
+        },
+        entity =
+        {
+            settings =
+            {
+                ["big-rock"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["big-sand-rock"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                coal =
+                {
+                    frequency = 0,
+                    richness = 0,
+                    size = 0
+                },
+                ["copper-ore"] =
+                {
+                    frequency = 0,
+                    richness = 0,
+                    size = 0
+                },
+                ["crude-oil"] =
+                {
+                    frequency = 0,
+                    richness = 0,
+                    size = 0
+                },
+                fish =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["huge-rock"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["iron-ore"] =
+                {
+                    frequency = 0,
+                    richness = 0,
+                    size = 0
+                },
+                stone =
+                {
+                    frequency = 0,
+                    richness = 0,
+                    size = 0
+                },
+                ["uranium-ore"] =
+                {
+                    frequency = 0,
+                    richness = 0,
+                    size = 0
+                }
+            },
+            treat_missing_as_default = true
+        },
+        tile =
+        {
+            settings =
+            {
+                ["gleba-deep-lake"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["highland-dark-rock"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["highland-dark-rock-2"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["highland-yellow-rock"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["lowland-brown-blubber"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["lowland-cream-cauliflower"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["lowland-cream-cauliflower-2"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["lowland-cream-red"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["lowland-dead-skin"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["lowland-dead-skin-2"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["lowland-olive-blubber"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["lowland-olive-blubber-2"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["lowland-olive-blubber-3"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["lowland-pale-green"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["lowland-red-infection"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["lowland-red-vein"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["lowland-red-vein-2"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["lowland-red-vein-3"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["lowland-red-vein-4"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["lowland-red-vein-dead"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["midland-cracked-lichen"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["midland-cracked-lichen-dark"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["midland-cracked-lichen-dull"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["midland-turquoise-bark"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["midland-turquoise-bark-2"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["midland-yellow-crust"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["midland-yellow-crust-2"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["midland-yellow-crust-3"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["midland-yellow-crust-4"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["natural-jellynut-soil"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["natural-yumako-soil"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["pit-rock"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["wetland-blue-slime"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["wetland-dead-skin"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["wetland-green-slime"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["wetland-jellynut"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["wetland-light-dead-skin"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["wetland-light-green-slime"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["wetland-pink-tentacle"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["wetland-red-tentacle"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                },
+                ["wetland-yumako"] =
+                {
+                    frequency = 1,
+                    richness = 1,
+                    size = 1
+                }
+            },
+            treat_missing_as_default = true
+        }
+    },
+    cliff_settings =
+    {
+        cliff_elevation_0 = 0,
+        cliff_elevation_interval = 0,
+        cliff_smoothing = 0,
+        control = "nauvis_cliff",
+        name = "cliff",
+        richness = 1
+    },
+    default_enable_all_autoplace_controls = false,
+    no_enemies_mode = false,
+    peaceful_mode = false,
+    property_expression_names = {},
+    starting_area = 10,
+    starting_points =
+    {
+        {
+            x = 0,
+            y = 0
+        }
+    },
 }
 
 return map_gen_settings
