+++ maps/infestation_islands/main.lua
@@ -1,14 +1,14 @@
 --created by Gerkiz
 local Public = require 'maps.infestation_islands.core'
 local Event = require 'utils.event'
-local Func = Public.func
-local Task = require 'utils.task_token'
+local Func = Public.functions
 local Scheduler = require 'utils.scheduler'
 local Difficulty = require 'modules.difficulty_vote_by_amount'
 local Server = require 'utils.server'
 local Gui = require 'utils.gui'
 
 local stage_gui_name = Gui.uid()
+local random = math.random
 
 if not script.active_mods.quality then
     error('Quality mod is not enabled!')
@@ -48,7 +48,8 @@ local function reset_player(player)
 end
 
 local reset_players_token =
-    Task.register(
+    Scheduler.register_function(
+        'reset_players_token',
         function ()
             local surface = game.get_surface(1)
 
@@ -82,69 +83,82 @@ end
 
 local function create_stage_gui(player)
     local button = get_top_frame(player, stage_gui_name)
-    if button then
-        return
-    end
+    if button and button.valid then button.destroy() end
 
     if Gui.get_mod_gui_top_frame() then
-        local frame =
+        local element =
             Gui.add_mod_button(
                 player,
                 {
                     type = 'frame',
                     name = stage_gui_name,
-                    caption = ' '
                 }
             )
-        if frame then
-            frame.style.minimal_height = 36
-            frame.style.maximal_height = 36
-            frame.style.minimal_width = 140
-            frame.style.maximal_width = 420
-            frame.style.font_color = { r = 155, g = 85, b = 25 }
-            frame.style.font = 'heading-2'
+        if element and element.valid then
+            local style = element.style
+
+            style.minimal_height = 36
+            style.maximal_height = 36
+            style.padding = 0
+            local label = element.add({ type = 'label', caption = ' ', name = 'label' })
+            label.style.font_color = { r = 0.88, g = 0.88, b = 0.88 }
+            label.style.font = 'heading-1'
         end
     else
         local element = player.gui.top.add({ type = 'frame', name = stage_gui_name, caption = ' ' })
-        local style = element.style
-        style.minimal_height = 54
-        style.maximal_height = 54
-        style.minimal_width = 140
-        style.maximal_width = 420
-        style.top_padding = 12
-        style.left_padding = 4
-        style.right_padding = 4
-        style.bottom_padding = 2
-        style.font_color = { r = 155, g = 85, b = 25 }
-        style.font = 'default-large-bold'
+        if element and element.valid then
+            local style = element.style
+            style.padding = 0
+
+            local label = element.add({ type = 'label', caption = ' ', name = 'label' })
+            label.style.font_color = { r = 0.88, g = 0.88, b = 0.88 }
+            label.style.font = 'heading-1'
+        end
     end
 end
 
-local function update_stage_gui(caption_override)
+local function update_stage_gui()
     local this = Public.get()
     if not this.stages then
         return
     end
-    local caption = 'Level: ' .. this.current_level
-    caption = caption .. '  |  Stage: '
-    local stage = this.current_stage
+    local caption = 'Level: '
+    local stage = this.current_level
     if stage > #this.stages - 1 then
         stage = #this.stages - 1
     end
+    local islands_data = Public.get('islands_data')
+    local island_data = islands_data[stage]
     caption = caption .. stage
     caption = caption .. '/'
     caption = caption .. #this.stages - 1
-    if this.alive_enemies == 0 then
-        caption = caption .. '  |  Level cleared!'
-    else
-        caption = caption .. '  |  Bugs remaining: '
+    local tooltip
+    if not (island_data and island_data.spawned_biters) then
+        caption = caption .. ' | Generating...'
+        tooltip = 'The biters are still generating on the island. Please wait for them to finish.'
+    elseif (this.auto_generate_upon_idle and island_data and island_data.spawned_biters and island_data.completed and this.time_until_next_island_is_created and this.time_until_next_island_is_created > game.tick and Difficulty.has_votes_ended()) then
+        caption = caption .. ' | Level cleared!'
+        caption = caption .. ' | [entity=small-biter,quality=' .. Public.qualities[random(1, #Public.qualities)] .. ']: ' .. Public.normalize_time_until_next_island_is_created()
+        tooltip = 'The next island will be generated in ' .. Public.normalize_time_until_next_island_is_created() .. '.\nUnless you progress to the next island, it will be generated automatically.\nIf you do not progress to the next island, you will not be able to reroll the next island market.\nMarket rerolls are unlocked when you manually progress to the next island.'
+    elseif (island_data and island_data.spawned_biters and island_data.completed) then
+        caption = caption .. ' | Level cleared!'
+        tooltip = 'Defenses would sure be helpful right now.\nVotes close in ' .. Difficulty.get_closing_timeout() .. ' seconds.'
+    elseif (this.auto_generate_upon_idle and island_data and island_data.auto_generated_bridge == false) then
+        caption = caption .. ' | Bugs remaining: '
         caption = caption .. this.alive_enemies
+        caption = caption .. ' | [entity=small-biter,quality=' .. Public.qualities[random(1, #Public.qualities)] .. ']: ' .. Public.normalize_time_until_next_island_is_created()
+        tooltip = 'The bridge to the next island will be generated in ' .. Public.normalize_time_until_next_island_is_created() .. '.\nUnless you progress to the next island, it will be generated automatically.\nMarket rerolls have been removed for the next island.\nMarket rerolls are unlocked when you manually progress to the next island.'
+    else
+        caption = caption .. ' | Bugs remaining: '
+        caption = caption .. this.alive_enemies
+        tooltip = 'Vanquish the biters to capture the island. ' .. this.alive_enemies .. ' biters remaining.'
     end
 
     for _, player in pairs(game.connected_players) do
         local frame = get_top_frame(player, stage_gui_name)
-        if frame then
-            frame.caption = caption_override or caption
+        if frame and frame.valid and frame.label and frame.label.valid then
+            frame.label.caption = this.top_label_caption_override or caption
+            frame.label.tooltip = tooltip
         end
     end
 end
@@ -172,6 +186,16 @@ end
 local function clear_surface()
     local surface = game.get_surface(1)
     surface.clear()
+
+    local planet = game.planets['nauvis']
+    local platforms = planet.get_space_platforms('player')
+    if platforms then
+        for _, platform in pairs(platforms) do
+            if platform and platform.valid then
+                platform.destroy()
+            end
+        end
+    end
 end
 
 local function on_player_joined_game(event)
@@ -195,7 +219,9 @@ local function has_the_game_ended(this)
         end
 
         game.forces.enemy.set_friend('player', true)
+        game.forces.enemy.set_cease_fire('player', true)
         game.forces.player.set_friend('enemy', true)
+        game.forces.player.set_cease_fire('enemy', true)
 
         this.game_reset_tick = this.game_reset_tick - 1
         if this.game_reset_tick % 600 == 0 then
@@ -224,7 +250,7 @@ local function has_the_game_ended(this)
                     this.ammo_chest = nil
                 end
 
-                for _, market_data in pairs(this.spawned_markets) do
+                for _, market_data in pairs(this.islands_data) do
                     if market_data and market_data.market and market_data.market.valid then
                         if market_data.render_protect_text then
                             market_data.render_protect_text.destroy()
@@ -235,16 +261,17 @@ local function has_the_game_ended(this)
                             market_data.render_checkpoint_text = nil
                         end
                         market_data.market.destroy()
+                        market_data.market = nil
                     end
                 end
 
                 this.game_reset_tick = nil
                 this.game_lost = false
                 this.game_won = false
-                Scheduler.can_run_scheduler(false)
+                Scheduler.clear_tasks()
                 clear_surface()
                 Public.on_init()
-                Task.set_timeout_in_ticks(500, reset_players_token)
+                Scheduler.new(500, reset_players_token)
                 return
             end
 
@@ -274,27 +301,83 @@ end
 
 local function on_tick()
     local this = Public.get()
-    if game.tick % 25 == 0 then
-        if this.alive_enemies < 0 then this.alive_enemies = 0 end
-        if this.game_lost then
-            local message = this.nomed_marked and 'The bugs had a feast on the marked at level ' .. this.nomed_marked .. '!' or 'The bugs had a feast on the marked!'
-            update_stage_gui(message)
-        else
-            update_stage_gui()
-        end
-        if not this.game_lost then
-            if Difficulty.has_votes_ended() and not this.difficulty_vote_ended then
-                this.difficulty_vote_ended = true
-                game.print('The difficulty vote has ended! You may now progress to the next island!', { color = { r = 0.22, g = 0.88, b = 0.22 } })
-                Server.to_discord_embed('** The difficulty vote has ended! You may now progress to the next island! **')
-                game.print('The difficulty is ' .. Difficulty.get('name') .. '!', { color = Difficulty.get('print_color') })
-                Server.to_discord_embed('** The difficulty is ' .. Difficulty.get('name') .. '! **')
+    local island_data = this.islands_data[this.current_level]
+    local tick = game.tick
+
+    if this.delayed_messages[tick] then
+        game.print(this.delayed_messages[tick])
+        this.delayed_messages[tick] = nil
+    end
+
+    if tick % 25 == 0 then
+        Func.check_alive_enemies()
+    end
+
+    has_the_game_ended(this)
+
+    if tick % 40 == 0 and this.auto_generate_upon_idle then
+        if island_data and Difficulty.has_votes_ended() then
+            local _, time = Public.normalize_time_until_next_island_is_created()
+
+            if island_data.completed then
+                if not this.time_until_next_island_is_created then
+                    local difficulty_index = Difficulty.get('index')
+                    local hour
+                    if difficulty_index == 1 then
+                        hour = random(60, 120)
+                    elseif difficulty_index == 2 then
+                        hour = random(30, 60)
+                    elseif difficulty_index == 3 then
+                        hour = random(15, 30)
+                    end
+
+                    this.time_until_next_island_is_created = tick + (60 * 60 * hour * this.current_level)
+                    this.time_until_next_island_is_created_static = math.round((this.time_until_next_island_is_created - tick) / 60 / 60, 0)
+
+                    if _DEBUG then
+                        this.time_until_next_island_is_created = tick + (60 * 60 * 10)
+                        this.time_until_next_island_is_created_static = math.round((this.time_until_next_island_is_created - tick) / 60 / 60, 0)
+                    end
+                    return
+                end
+
+                local time_limit = this.time_until_next_island_is_created_static / 2
+
+                -- spawn the island before the time limit is reached
+                if time <= time_limit then
+                    if not island_data.auto_generated_island then
+                        island_data.auto_generated_island = true
+                        game.print('The biters are getting hungry!!!', { color = { r = 0.88, g = 0.22, b = 0.22 } })
+                        Scheduler.new(1, Public.init_next_island_without_bridge_token):set_data({ surface = game.surfaces[1] })
+                    end
+                end
+            else
+                -- spawn the island after the time limit is reached
+                if time <= 0 then
+                    if not (island_data and island_data.auto_generated_bridge) then
+                        island_data.auto_generated_bridge = true
+                        game.print('The biters are forming a bridge to our island! They are coming!!!', { color = { r = 0.88, g = 0.22, b = 0.22 } })
+                        Scheduler.new(1, Public.do_generate_bridge_token):set_data({ surface = game.surfaces[1] })
+                        this.time_until_next_island_is_created = nil
+                    end
+                end
             end
         end
     end
 
-    local infinite_ammo_tick = Public.get('infinite_ammo_tick')
-    if game.tick % infinite_ammo_tick == 0 then
+    if tick % 50 == 0 then
+        update_stage_gui()
+        if this.game_lost then return end
+
+        local position = island_data and island_data.position or { x = 0, y = 0 }
+        local radius = island_data and island_data.radius or 0
+
+        game.forces.player.chart(game.surfaces[1], { { position.x - radius, position.y - radius }, { position.x + radius, position.y + radius } })
+
+        Func.is_rocket_silo_alive()
+    end
+
+    if tick % this.infinite_ammo_tick == 0 then
         drift_corpses_toward_beach()
         if this.ammo_chest and this.ammo_chest.valid then
             local magazine_name = 'firearm-magazine'
@@ -309,60 +392,45 @@ local function on_tick()
         end
     end
 
-    if game.tick % 150 == 0 then
-        if this.game_lost then return end
-
-        Func.is_rocket_silo_alive()
-
-        local center_position = this.centered_points[this.current_level]
-        if not center_position then
-            center_position =
-            {
-                position = { x = 0, y = 0 }
-            }
-        end
-        game.forces.player.chart(game.surfaces[1], { { center_position.position.x - 124, center_position.position.y - 124 }, { center_position.position.x + 124, center_position.position.y + 124 } })
-
-        Func.check_alive_enemies()
-        if not this.completed_levels[this.current_level] then
-            if not this.disable_multi_command_attack then
-                Func.do_buried_biters()
+    if tick % 100 == 0 then
+        if not this.game_lost then
+            if Difficulty.has_votes_ended() and not this.difficulty_vote_ended then
+                this.difficulty_vote_ended = true
+                game.print('The difficulty vote has ended! You may now progress to the next island!', { color = { r = 0.22, g = 0.88, b = 0.22 } })
+                Server.to_discord_embed('** The difficulty vote has ended! You may now progress to the next island! **')
+                game.print('The difficulty is ' .. Difficulty.get('name') .. '!', { color = Difficulty.get('print_color') })
+                Server.to_discord_embed('** The difficulty is ' .. Difficulty.get('name') .. '! **')
             end
         end
     end
 
-    if game.tick % 400 == 0 then
+    if tick % 150 == 0 then
+        if not this.game_lost then
+            if not island_data.completed then
+                if not this.disable_multi_command_attack then
+                    Func.do_buried_biters()
+                end
+            end
+        end
+    end
+
+    if tick % 400 == 0 then
         Func.set_multi_command()
     end
 
-    if game.tick % 500 == 0 then
-        if this.completed_levels[this.current_level] then
-            if not this.disable_multi_command_attack then
-                Func.do_buried_biters_on_completed_levels()
-            end
-        end
-    end
-
-    if this.delayed_messages[game.tick] then
-        game.print(this.delayed_messages[game.tick])
-        this.delayed_messages[game.tick] = nil
-    end
-
-    if game.tick % 500 == 0 then
+    if tick % 500 == 0 then
         Func.update_evolution_static()
     end
 
     if this.clear_items_on_ground_state then
-        if game.tick % 450 == 0 then
+        if tick % 450 == 0 then
             Func.do_clear_items_on_ground_slowly()
         end
 
-        if game.tick % 4500 == 0 then
+        if tick % 4500 == 0 then
             Func.run_clear_items_on_ground()
         end
     end
-
-    has_the_game_ended(this)
 end
 
 local handle_changes = function ()
